+++ bytecode/peephole_opt.py
@@ -125,8 +125,7 @@ class PeepholeOptimizer:
             return None
         if next_instr.name == 'POP_JUMP_IF_FALSE':
             # Replace UNARY_NOT+POP_JUMP_IF_FALSE with POP_JUMP_IF_TRUE
-            # FIXME: don't repeat lineno
-            instr.set('POP_JUMP_IF_TRUE', next_instr.arg, lineno=instr.lineno)
+            instr.set('POP_JUMP_IF_TRUE', next_instr.arg)
             self.block[self.index-1:self.index+1] = (instr,)
             self.index -= 1