+++ doc/changelog.rst
@@ -4,6 +4,7 @@ ChangeLog
 * Version 0.3
 
   - :class:`Instr` now validates the type of the argument
+  - Remove *lineno* parameter from :meth:`Instr.set`
   - add :class:`CellVar` and :class:`FreeVar` classes
   - peephole_opt: bugfixes
   - :class:`BytecodeBlocks`: remove undocumented ``to_concrete_bytecode()``